chore: fix the error-log list as the system route
