+++ src/enums/pageEnum.ts
@@ -6,5 +6,5 @@ export enum PageEnum {
   // error page path
   ERROR_PAGE = '/exception',
   // error log page path
-  ERROR_LOG_PAGE = '/feat/error-log',
+  ERROR_LOG_PAGE = '/error-log/list',
 }
