+++ src/layouts/default/header/components/index.ts
@@ -10,6 +10,4 @@ export const FullScreen = createAsyncComponent(() => import('./FullScreen.vue'))
 
 export const Notify = createAsyncComponent(() => import('./notify/index.vue'));
 
-export const LockItem = createAsyncComponent(() => import('./lock/index.vue'));
-
 export const ErrorAction = createAsyncComponent(() => import('./ErrorAction.vue'));
