--- src/layouts/default/header/components/lock/LockAction.less
@@ -1,30 +1,0 @@
-.lock-modal {
-  &__entry {
-    position: relative;
-    height: 240px;
-    padding: 130px 30px 60px 30px;
-    background: #fff;
-    border-radius: 10px;
-  }
-
-  &__header {
-    position: absolute;
-    top: 0;
-    left: calc(50% - 45px);
-    width: auto;
-    text-align: center;
-
-    &-img {
-      width: 70px;
-      border-radius: 50%;
-    }
-
-    &-name {
-      margin-top: 5px;
-    }
-  }
-
-  &__footer {
-    text-align: center;
-  }
-}
