--- src/layouts/default/menu/index.less
@@ -1,22 +1,0 @@
-@prefix-cls: ~'@{namespace}-layout-menu';
-@logo-prefix-cls: ~'@{namespace}-app-logo';
-
-.@{prefix-cls} {
-  &-logo {
-    height: @header-height;
-    padding: 10px 4px 10px 10px;
-
-    img {
-      width: @logo-width;
-      height: @logo-width;
-    }
-  }
-
-  &--mobile {
-    .@{logo-prefix-cls} {
-      &__title {
-        opacity: 1;
-      }
-    }
-  }
-}
