+++ src/layouts/default/setting/index.vue
@@ -1,20 +1,20 @@
 <template>
   <div @click="openDrawer" :class="prefixCls">
-    <SettingOutlined />
+    <Icon icon="ion:settings-outline" />
     <SettingDrawer @register="register" />
   </div>
 </template>
 <script lang="ts">
   import { defineComponent } from 'vue';
-  import { SettingOutlined } from '@ant-design/icons-vue';
   import SettingDrawer from './SettingDrawer';
+  import Icon from '/@/components/Icon';
 
   import { useDrawer } from '/@/components/Drawer';
   import { useDesign } from '/@/hooks/web/useDesign';
 
   export default defineComponent({
     name: 'SettingButton',
-    components: { SettingOutlined, SettingDrawer },
+    components: { SettingDrawer, Icon },
     setup() {
       const [register, { openDrawer }] = useDrawer();
