+++ src/layouts/default/sider/index.vue
@@ -17,8 +17,9 @@
   import { defineComponent } from 'vue';
 
   import Sider from './LayoutSider.vue';
-  import { Drawer } from 'ant-design-vue';
   import MixSider from './MixSider.vue';
+  import { Drawer } from 'ant-design-vue';
+
   import { useAppInject } from '/@/hooks/web/useAppInject';
   import { useMenuSetting } from '/@/hooks/setting/useMenuSetting';
   import { useDesign } from '/@/hooks/web/useDesign';
