+++ src/locales/lang/en/routes/basic.ts
@@ -1,3 +1,4 @@
 export default {
   login: 'Login',
+  errorLogList: 'Error Log',
 };
