+++ src/locales/lang/zh_CN/routes/basic.ts
@@ -1,3 +1,4 @@
 export default {
   login: '登录',
+  errorLogList: '错误日志列表',
 };
