+++ src/router/constant.ts
@@ -1,5 +1,6 @@
 import type { AppRouteRecordRaw } from '/@/router/types';
 import ParentLayout from '/@/layouts/page/ParentView.vue';
+import { t } from '/@/hooks/web/useI18n';
 
 const EXCEPTION_COMPONENT = () => import('../views/sys/exception/Exception.vue');
 
@@ -60,6 +61,27 @@ export const REDIRECT_ROUTE: AppRouteRecordRaw = {
       component: () => import('/@/views/sys/redirect/index.vue'),
       meta: {
         title: REDIRECT_NAME,
+        hideBreadcrumb: true,
+      },
+    },
+  ],
+};
+
+export const ERROR_LOG_ROUTE: AppRouteRecordRaw = {
+  path: '/error-log',
+  name: 'errorLog',
+  component: LAYOUT,
+  meta: {
+    title: 'ErrorLog',
+    hideBreadcrumb: true,
+  },
+  children: [
+    {
+      path: 'list',
+      name: 'errorLogList',
+      component: () => import('/@/views/sys/error-log/index.vue'),
+      meta: {
+        title: t('routes.basic.errorLogList'),
         hideBreadcrumb: true,
       },
     },
